pyproject.toml
[project]
name = "defenseextraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "selenium>=4.39.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "xlsxwriter>=3.2.9",
]
